+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - `greetings.yml` removed.
+- `test.py` modified.
 ## [5.3] - 2021-08-30
 ### Added
 - 10 new font