+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- 7 new font
+- 10 new font
 	1. fancy131
 	2. fancy132
 	3. fancy133
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	5. fancy135
 	6. fancy136
 	7. fancy137
+	8. tarty1
+	9. tarty2
+	10. tarty3
 - Discord badge
 ### Changed
 - `test.py` modified