+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `test.py` modified
 - `font_wizard.py` modified
 - `font_check.py` modified
+- `art_decor_check.py` modified
 - `__word2art` function modified
 - `mix_letters` function modified
 - `AUTHORS.md` updated